moving common cli and logging out of mpx
